# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DeliveryApp API",
        "description": "API for delivery management system",
        "contact": {
            "email": "admin@deliveryapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/client/requests": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "List client's delivery requests",
                "operationId": "79d0fc5ae68a4d395046d37777557e6d",
                "responses": {
                    "200": {
                        "description": "List of requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "requests": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Create a new delivery request",
                "operationId": "ce8daa161e35a9604b5761dccfb7bba2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item",
                                    "pickup_location",
                                    "dropoff_location"
                                ],
                                "properties": {
                                    "item": {
                                        "required": [
                                            "type"
                                        ],
                                        "properties": {
                                            "type": {
                                                "type": "string",
                                                "example": "Documento"
                                            },
                                            "weight_kg": {
                                                "type": "number",
                                                "example": 0.5
                                            },
                                            "height_cm": {
                                                "type": "number",
                                                "example": 30
                                            },
                                            "width_cm": {
                                                "type": "number",
                                                "example": 20
                                            },
                                            "length_cm": {
                                                "type": "number",
                                                "example": 5
                                            },
                                            "notes": {
                                                "type": "string",
                                                "example": "Frágil"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "pickup_location": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "example": -23.55052
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "example": -46.633308
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "dropoff_location": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "example": -23.561414
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "example": -46.656166
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created"
                    },
                    "400": {
                        "description": "Client already has active request"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/client/requests/{id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Get request details",
                "operationId": "055d85eadad94434a60f831511b4c796",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request details"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/client/requests/{id}/cancel": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Cancel a request",
                "operationId": "7481b5e2e800ba81cb282fa2b331d864",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request canceled"
                    },
                    "400": {
                        "description": "Cannot cancel this request"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/client/requests/{id}/share": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Generate share link for tracking",
                "operationId": "67994f81fa76b51e761c67165834cab7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Share link generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "share_token": {
                                            "type": "string"
                                        },
                                        "share_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot share this request"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/client/requests/{id}/tracking": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Get deliverer location for tracking",
                "operationId": "0074d7027ac189200c624a55664149e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deliverer location",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "latitude": {
                                            "type": "number"
                                        },
                                        "longitude": {
                                            "type": "number"
                                        },
                                        "updated_at": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No active delivery"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}